Move buildinfo to util
